+++ sop-admin/sop-admin-server/src/main/java/com/gitee/sop/adminserver/api/service/MonitorNewApi.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
+ * 路由监控
  * @author tanghc
  */
 @ApiService