+++ sop-admin/sop-admin-server/src/main/java/com/gitee/sop/adminserver/service/ConfigPushService.java
@@ -37,9 +37,6 @@ public class ConfigPushService {
     @Autowired
     private ServerService serverService;
 
-    @Value("${gateway.host:}")
-    private String gatewayHost;
-
     @Value("${sop.secret}")
     private String secret;