+++ sop-website/sop-website-server/src/main/java/com/gitee/sop/websiteserver/controller/result/DocVO.java
@@ -1,0 +1,17 @@
+package com.gitee.sop.websiteserver.controller.result;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Collection;
+
+/**
+ * @author tanghc
+ */
+@Getter
+@Setter
+public class DocVO {
+    private String urlTest;
+    private String urlProd;
+    private Collection<MenuProject> menuProjects;
+}